Module oxcable::chain

A container for a series of audio devices.

A chain can be used when a single series of audio devices passes its output to the input of the next device. It is initialized from a single starting device that will receive no input, and ends in a device who's output is ignored.

Example

The following will pass microphone input through a delay unit, then out to the speaker:

use oxcable::delay::Delay;
use oxcable::chain::{DeviceChain, Tick};
use oxcable::io::audio::AudioEngine;

let engine = AudioEngine::with_buffer_size(256).unwrap();
let mut chain = DeviceChain::from(
    engine.default_input(1).unwrap()
).into(
    Delay::new(1.0, 0.5, 0.5, 1)
).into(
    engine.default_output(1).unwrap()
);
chain.tick_forever();

Reexports

pub use tick::Tick;

Structs

DeviceChain
A container for a series of audio devices.