Struct oxcable::dynamics::level_detector::LevelDetector

pub struct LevelDetector {
    // some fields omitted
}

Performs envelope estimation for a signal.

The level detector performs leaky integration of the input signal to estimate the envelope. Therefore, for a signal x(t), the level l(t) = alpha*l(t-1) + (1-alpha)*x(t).

When creating a new level detector, a tau is specified, rather than an alpha. Tau is the time constant, which corresponds to how quickly the envelope estimate reaches steady state.

The integration uses two different values of tau, depending on whether the signal is greater or less than the current envelope. This means the attack value can be set very low to respond quickly to bursts in signal power, while the release value can be high in order to coast through periodic troughs in the signal.

Methods

impl LevelDetector

fn default() -> Self

Returns a level detector with default tau values tuned for reasonable performance.

fn new(attack_tau: f32, release_tau: f32) -> Self

Returns a new level detector.

attack_tau specifies the time constant when the signal is growing, in milliseconds.

release_tau specifies the time constant when the signal id decaying, in milliseconds.

fn compute_next_level(&mut self, s: Sample) -> f32

Given the next input sample s, computes the current estimate of the envelope value.