Struct oxcable::graph::DeviceGraph

pub struct DeviceGraph {
    // some fields omitted
}

An acyclic graph for audio devices.

Methods

impl DeviceGraph

fn new() -> Self

Creates an empty graph.

fn add_node<D>(&mut self, device: D) -> AudioNodeIdx where D: 'static + AudioDevice

Adds a new device into the graph, with no connections. Returns a identifier that refers back to this device.

fn add_edge(&mut self, src: AudioNodeIdx, src_ch: usize, dest: AudioNodeIdx, dest_ch: usize) -> Result<()>

Connects two devices in the graph.

src and dest are identifiers for the actual devices to connect.

src_ch and dest_ch are the channel indices of the two devices.

If invalid indices are provided, or if the specified edge would create a cycle in the graph, an Err is returned and no changes dest the graph are made.