Module oxcable::io::midi

MIDI input from system MIDI devices.

A MidiEngine is used to manage the MIDI driver and open new MIDI streams. All input streams must be opened through an engine instance.

Panics

The tick functions for both input streams will panic if they can't read from the system MIDI stream.

Structs

MidiEngine
A system resources manager.

MidiIn
Reads audio from the OS's default midi device.