Struct oxcable::io::wav::WavReader

pub struct WavReader<R: Read> {
    // some fields omitted
}

Reads audio from a wav file.

The reader will continue until it runs out of samples. When it does, the reader will return silence until it is reset to the beginning of the file.

Methods

impl WavReader<File>

fn open<P: AsRef<Path>>(filename: P) -> Result<Self>

Returns a WavReader reading the provided file.

impl<R: Read> WavReader<R>

fn new(reader: R) -> Result<Self>

Retuns a WavReader reading anything implementing Read.

fn get_num_samples(&self) -> Time

Returns the number of audio samples in the wav file.

fn is_done(&self) -> bool

Returns true if we have read the entire wav file.

impl<R: Read + Seek> WavReader<R>

fn restart(&mut self) -> Result<u64>

Resets the reader to begin reading from the start of the file.