Struct oxcable::io::wav::WavWriter

pub struct WavWriter<W: Write + Seek> {
    // some fields omitted
}

Writes audio to a wav file.

The writer initializes the data_size to be 0. This will not be overwritten with the proper size until update_data_size is called.

While WavReader only requires its type be Seek to use the restart method, the WavWriter reqiures Seek for all types, because the final data size must be written to the header when the writer leaves scope.

Methods

impl WavWriter<File>

fn create<P: AsRef<Path>>(filename: P, num_channels: usize) -> Result<Self>

Returns a WavWriter writing to the provided file.

impl<W: Write + Seek> WavWriter<W>

fn new(writer: W, num_channels: usize) -> Result<Self>

Returns a WavWriter writing to anything implementing Write.