Struct oxcable::reverb::moorer_reverb::MoorerReverb

pub struct MoorerReverb {
    // some fields omitted
}

An algorithmic, IIR reverberation filter.

This algorithmic reverb filter follows the basic design specified by James Moorer in his seminal paper, "About This Reverberation Business".

Methods

impl MoorerReverb

fn new(room: Room, rev_time: f32, gain: f32, wetness: f32, num_channels: usize) -> Self

Creates and initializes a new reverberator.

room: a Room struct. This specifies aspects of the reverberation character.

rev_time: the -60dB time, in seconds.

gain: the output gain, in decibels.

wetness: how much of the input signal to mix into the output.

num_channels: number of channels to process.