Trait oxcable::tick::Tick

pub trait Tick {
    fn tick(&mut self);

    fn tick_n_times(&mut self, n: Time) { ... }
    fn tick_forever(&mut self) -> ! { ... }
    fn tick_until_enter(&mut self) { ... }
}

Methods for processing something in discrete time steps. By defining how to perform a single tick, this trait gives several convenience methods.

A single tick could correspond to one frame, but it doesn't have to. For example, if an object does buffering it may process many samples per tick.

Required Methods

fn tick(&mut self)

Handles one time step.

Provided Methods

fn tick_n_times(&mut self, n: Time)

Runs tick n times. Returns after processing.

fn tick_forever(&mut self) -> !

Ticks into infinity. Never returns.

fn tick_until_enter(&mut self)

Ticks while waiting for the user to press Enter. When enter is pressed, ticking stops and the method returns.