Enum oxcable::types::MidiMessage

pub enum MidiMessage {
    NoteOn(u8, f32),
    NoteOff(u8, f32),
    PitchBend(f32),
    PolyphonicAftertouch(u8, f32),
    ChannelAftertouch(f32),
    SustainPedal(bool),
    Portamento(bool),
    Sostenuto(bool),
    SoftPedal(bool),
    LegatoPedal(bool),
    Hold2(bool),
    ControlChange(u8, u8),
    ProgramChange(u8),
    Other(u8, u8, u8),
}

The contents of a MIDI message.

This enum wraps a useful subset of possible MIDI message. For further details on what these message represent, consult a complete MIDI documentation.

Certain messages are parsed out to more useful datatypes:

Velocities are converted to floats between 0.0 and 1.0

Pressures are converted to floats between 0.0 and 1.0

Bend is converted to a float from -1.0 to 1.0

Any MIDI type not specifically enumerated preserves the raw bytes as Other.

Variants

NoteOn

NoteOn(note number, velocity)

NoteOff

NoteOff(note number, velocity)

PitchBend

PitchBend(bend)

PolyphonicAftertouch

PolyphonicAftertouch(note number, pressure)

ChannelAftertouch

ChannelAftertouch(pressure)

SustainPedal

SustainPedal(on/off)

Portamento

Portamento(on/off)

Sostenuto

Sostenuto(on/off)

SoftPedal

SoftPedal(on/off)

LegatoPedal

LegatoPedal(on/off)

Hold2

Hold2(on/off)

ControlChange

ControlChange(controller, value)

ProgramChange

ProgramChange(num)

Other

Other(status, byte1, byte2)