Struct oxcable::utils::fft::Transformer

pub struct Transformer {
    // some fields omitted
}

A container for precomputed values to perform FFTs of a fixed size.

Methods

impl Transformer

fn new(size: usize) -> Self

Initializes a transformer to perform FFTs of the provided size, rounded up to the nearest power of two.

fn get_size(&self) -> usize

Returns the size FFT this Transformer performs.

fn fft(&self, input: &[Complex32], output: &mut [Complex32])

Performs an FFT on input, and place the result in output.

The input is zero padded if less than size samples are provided, and truncated if more than size samples are provided.

fn ifft(&self, input: &[Complex32], output: &mut [Complex32])

Performs an inverse FFT on input, and place the result in output.

The input is zero padded if less than size samples are provided, and truncated if more than size samples are provided.