Struct oxcable::utils::ringbuffer::RingBuffer

pub struct RingBuffer<T> {
    // some fields omitted
}

A generic ring buffer.

A ring buffer can continue appending data to itself indefinitely. However, it has a limited capacity; when that capacity is reached, it will overwrite the oldest data.

The buffer is indexed with Time values. As data is overwritten, the times stored in the buffer continue to slide forward.

Initialization from slices

Sometimes, it helps to be able to specify the initial state of a ring buffer. One example of this is to create a silent buffer of audio that delayed samples can be accumulated in, such as may be used to generate delays.

To make this easier, a buffer can be created from a slice. When created from a slice, the buffer will be have a capacity that matches the size of the slice, and the values of the slice will be inserted into the buffer, starting at time t=0.

let values = ["first", "second", "third"];
let rb = RingBuffer::from(&values[..]);
assert_eq!(rb.capacity(), 3);
assert_eq!(rb[0], "first");
assert_eq!(rb[1], "second");
assert_eq!(rb[2], "third");

Example

// Create a new buffer that holds 4 elements...
let mut rb = RingBuffer::new(4);

// Insert 3 elements...
rb.push(1);
rb.push(2);
rb.push(3);
assert_eq!(rb.len(), 3);
assert_eq!(rb.get(0), Some(&1));
assert_eq!(rb.get(1), Some(&2));
assert_eq!(rb.get(2), Some(&3));

// Insert two more elements. This will push the oldest element out...
rb.push(4);
rb.push(5);
assert_eq!(rb.len(), 4);
assert_eq!(rb.get(0), None);
assert_eq!(rb.get(4), Some(&5));

// Pop the oldest element off the buffer...
assert_eq!(rb.get(1), Some(&2));
assert_eq!(rb.pop(), Some(2));
assert_eq!(rb.get(1), None);
assert_eq!(rb.len(), 3);

Methods

impl<T> RingBuffer<T>

fn new(capacity: usize) -> Self

Returns an empty ring buffer that can hold at most capacity elements.

fn len(&self) -> usize

Returns the number of elements the ringbuffer currently contains.

fn capacity(&self) -> usize

Returns the number of elements the ringbuffer can hold at one time.

fn times(&self) -> (Time, Time)

Returns the range of timestamps stored in the ringbuffer, as a tuple (first, last). First is inclusive, while last is exlusive.

Example

let mut rb = RingBuffer::new(2);
rb.push("first");  // t=0
rb.push("second"); // t=1
rb.push("third");  // t=2, first is overwritten
assert_eq!(rb.times(), (1,3));

fn get(&self, t: Time) -> Option<&T>

Attempts to return a ref to the data stored at time t. If the requested time is not in the buffer, instead returns None.

fn get_mut(&mut self, t: Time) -> Option<&mut T>

Attempts to return a mutable ref to the data stored at time t. If the requested time is not in the buffer, instead returns None.

fn push(&mut self, data: T)

Pushes the supplied data onto the end of the buffer. If the buffer is full, this will overwrite the oldest data.

fn clear(&mut self)

Clears all the elements from the buffer.

fn values(&self) -> Values<T>

Returns an iterator than returns the buffer's values.

Example

let mut rb = RingBuffer::new(2);
rb.push("first");
rb.push("second");
rb.push("third");

let values: Vec<_> = rb.values().collect();
assert_eq!(values, [&"second", &"third"]);

fn iter(&self) -> Iter<T>

Returns an iterator that pairs values with their timestamp.

Example

let mut rb = RingBuffer::new(2);
rb.push("first");
rb.push("second");
rb.push("third");

let zipped: Vec<_> = rb.iter().collect();
assert_eq!(zipped, [(1, &"second"), (2, &"third")]);

impl<T: Clone> RingBuffer<T>

fn pop(&mut self) -> Option<T>

Removes the oldest element from the ringbuffer and return it. If there are no elements, returns None insteads.

fn resize(&mut self, capacity: usize)

Resizes the ringbuffer to hold up to capacity elements. If the new capacity is smaller than the old one, then the oldest elements will be removed from the buffer.

Example

let mut rb = RingBuffer::new(4);
rb.push(1); rb.push(2); rb.push(3); rb.push(4);
rb.resize(2);
assert_eq!(rb.values().cloned().collect::<Vec<_>>(), [3,4]);