Struct oxcable::voice_array::VoiceArray

pub struct VoiceArray<T> {
    // some fields omitted
}

A manager for a polyphonic set of voices.

Methods

impl<T> VoiceArray<T>

fn new(voices: Vec<T>) -> Self

Creates a new VoiceArray from the provided vector of voices.

voices is moved inside the array.

fn iter(&self) -> Iter<T>

Returns an iterator over the voice objects.

fn iter_mut(&mut self) -> IterMut<T>

Returns a mutable iterator over the voice objects.

fn note_on(&mut self, note: u8) -> &mut T

Selects a new voice, marks it as playing, and loans it out for modification.

fn note_off(&mut self, note: u8) -> Option<&mut T>

Finds the voice playing the provided note, marks it as free, then loans it out for modification. If no voice is playing the provided note, then None is returned instead.