Struct oxcable::wrappers::BufferedOutput

pub struct BufferedOutput<D> where D: AudioDevice {
    pub device: D,
    pub outputs: Vec<Sample>,
}

Bundles an AudioDevice with an allocated output buffer.

To use the device, input samples are passed to tick. The output samples can be found in the outputs buffer.

Example

use oxcable::types::{AudioDevice, Sample, Time};
struct IdentityFilter;
impl AudioDevice for IdentityFilter {
    fn num_inputs(&self) -> usize { 1 }
    fn num_outputs(&self) -> usize { 1 }
    fn tick(&mut self, _: Time, inputs: &[Sample], outputs: &mut[Sample]) {
        outputs[0] = inputs[0];
    }
}

use oxcable::wrappers::BufferedOutput;
let mut filter = BufferedOutput::from(IdentityFilter);

for i in 0..8 {
    let input = [i as f32];
    filter.tick(i, &input);
    assert_eq!(i as f32, filter.outputs[0]);
}

Fields

device

The AudioDevice being wrapped.

outputs

The output buffer.

Methods

impl<D> BufferedOutput<D> where D: AudioDevice

fn tick(&mut self, t: Time, inputs: &[Sample])

Calls the device's tick method using the wrapper's buffers.